Massive performance improvements due to reduced $digest cycles
After throwing the component into a huge production application I
noticed that browsing through the select options didn't feel snappy
anymore. The reason was the excessive use of Angular bindings inside of
the directive. Since the only necessary integration point with Angular
is the ngModel update, I decided to use vanilla JavaScript event
listeners and replaced ngRepeat with my own implementation.

The result is a massive boost in performance.